refactor: rename to tasks_fn
